akher akher akher wahed

+++ src/deletet.cpp
@@ -7,6 +7,7 @@
 #include <QLineEdit>
 #include <QDebug>
 #include<graph.h>
+#include<QMessageBox>
 
 
 
@@ -28,6 +29,20 @@ void deletet::on_pushButton_clicked()
     QString sourceT = ui->lineEdit->text();
     QString destinationT = ui->lineEdit_2->text();
     QString Transportation = ui->lineEdit_3->text();
+    if (sourceT.isEmpty() || destinationT.isEmpty() || Transportation.isEmpty()) {
+        QMessageBox::critical(this, "Error", "Please fill in all fields.");
+        return;
+    }
+
+    // Check if both source and destination nodes are valid in the graph
+    if (!graph->isValidNode(sourceT) || !graph->isValidNode(destinationT)) {
+        QMessageBox::critical(this, "Error", "Invalid source or destination node.");
+        return;
+    }
+    if (!graph->isValidEdge(sourceT,destinationT,Transportation)) {
+        QMessageBox::critical(this, "Error", "Invalid transportation");
+        return;
+    }
     graph->deleteEdge(sourceT,destinationT,Transportation);
